macosx: Fix tiny bugs in volume control
Fix a problem in the parsing string used for the volume control tooltip.
The old string formatted a float as an int, leading to printing wrong
values.

Additionally, set the default value of the volume on startup to 1 (the
correct value), instead of 0.

Clean the LibraryWindow xib from 3 old variables, that were removed from
the class, but not the xib.

+++ modules/gui/macosx/windows/mainwindow/VLCMainWindowControlsBar.m
@@ -129,10 +129,10 @@
     BOOL b_muted = _playerController.mute;
 
     if (b_muted)
-        f_volume = 0.;
+        f_volume = 0.f;
 
     [self.volumeSlider setFloatValue: f_volume];
-    NSString *volumeTooltip = [NSString stringWithFormat:_NS("Volume: %i %%"), f_volume * 100];
+    NSString *volumeTooltip = [NSString stringWithFormat:_NS("Volume: %i %%"), (int)(f_volume * 100.0f)];
     [self.volumeSlider setToolTip:volumeTooltip];
 
     [self.volumeSlider setEnabled: !b_muted];
